修改了页面bug

+++ src/main/webapp/WEB-INF/views/admin/log/list.jsp
@@ -82,12 +82,16 @@
 	}
 	
 	function deleteLog(){
-		$.fn.jalert("您确定要删除该日志？",function(){
-			document.filterForm.method = "post";
-			document.filterForm.action = '${ctx}/admin/log/deleteAll?fdType=${fdType}&key='+$("#fdKey").val();
-			document.filterForm.submit();
-		});
-		
+		var str=$("input[name='ids']:checked").length;
+	    if(str==0){
+	    	$.fn.jalert("请选择用户");
+	    }else{
+	    	$.fn.jalert("您确定要删除该日志？",function(){
+				document.filterForm.method = "post";
+				document.filterForm.action = '${ctx}/admin/log/deleteAll?fdType=${fdType}&key='+$("#fdKey").val();
+				document.filterForm.submit();
+			});
+	    }
 	}
 </script>
 </head>
